requirements.txt
colorama==0.4.6
python-dotenv==1.1.0
openai==1.78.1
uvicorn==0.34.0
mcp==1.6.0
langchain==0.3.25
langchain-community==0.3.24
numpy==2.2.5
ollama==0.4.8
openai-agents==0.0.14
fastapi==0.115.9
pymetasploit3==1.0.6
tiktoken==0.9.0
PyYAML==6.0.2
# Add other necessary dependencies